+++ noodles-cram/CHANGELOG.md
@@ -6,7 +6,7 @@
 
   * cram/codecs: Add fqzcomp encoder.
 
-  * cram/codecs/name_tokenizer/encode: Add `DIGITS0`, `DIGITS`,
+  * cram/codecs/name_tokenizer/encode: Add `DIGITS0`, `DIGITS`, `DELTA`,
     `MATCH` tokens.
 
 ### Fixed